requirements.txt
faiss-cpu
sentence-transformers
openai
langchain
python-dotenv
huggingface-hub
google-generativeai
